Create AumentoSalario.java
(Switch) Write a program that takes the code corresponding to an employee's position and their current salary as input. The program should then determine and display the employee's position, the amount of the salary increase, and the new salary. The positions are listed in the table below:

CODE POSITION PERCENTAGE
1 Clerk 50%
2 Secretary 35%
3 Cashier 20%
4 Manager 10%
5 Director No increase